+++ tree-sitter-axiom/src/parser.c
@@ -40635,7 +40635,7 @@ TS_PUBLIC const TSLanguage *tree_sitter_axiom(void) {
     .metadata = {
       .major_version = 0,
       .minor_version = 7,
-      .patch_version = 4,
+      .patch_version = 5,
     },
   };
   return &language;